requirements.txt
Django==1.9.13
django-betterforms==1.1.4
django-ajax-selects==1.4.1
django-mptt==0.8.6
django-post-office==3.0.4
django-timezone-field==2.0
djangorestframework==3.3.2
gdata==2.0.18
httplib2==0.9.2
jsonfield==1.0.3
Markdown==2.6.5
oauth2client==1.5.1
psycopg2==2.7
pyasn1==0.1.9
pyasn1-modules==0.0.8
python-dateutil==2.4.2
pytz==2017.3
rsa==3.2
six==1.9.0
requests==2.18.4
django-paypal==0.5.0
pika==1.1.0
